backend/requirements.txt
anthropic
chromadb
sentence-transformers<3.0
pypdf
python-dotenv
python-docx
torch==2.2.2
numpy<2
fastapi
uvicorn
python-multipart
